+++ Campaign/database/migrations/2023_09_12_121415_create_campaign_collections_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateCampaignCollectionsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('campaign_collections', function (Blueprint $table) {
+            $table->id();
+            $table->bigInteger('collection_id')->unsigned();
+            $table->foreign('collection_id')->references('id')->on('collections');
+            $table->integer('campaign_id')->unsigned();
+            $table->foreign('campaign_id')->references('id')->on('campaigns');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('campaign_collections');
+    }
+}