+++ Campaign/resources/assets/js/app.js
@@ -10,5 +10,6 @@ require("@remp/js-commons/js/search.js");
 require("./banner.js");
 require("./forms/bannerForm.js");
 require("./forms/campaignForm.js");
+require("./forms/collectionForm.js");
 require("./forms/snippetForm.js");
 require("./javascript-twig-lint.js");