+++ Campaign/resources/views/schedule/index.blade.php
@@ -14,7 +14,7 @@
                 <div class="card-header">
                     <h2>Schedule <small></small></h2>
                     <div class="actions">
-                        <a href="{{ route('campaigns.index') }}" class="btn palette-Cyan bg waves-effect"><i class="zmdi zmdi-long-arrow-return"></i> Back to campaigns</a>
+                        <a href="{{ route('campaigns.index', ['collection' => $collection]) }}" class="btn palette-Cyan bg waves-effect"><i class="zmdi zmdi-long-arrow-return"></i> Back to campaigns</a>
                     </div>
                 </div>